requirements.txt
streamlit
pytesseract
pillow
opencv-python
numpy
PyMuPDF